+++ projects/pastanaga-angular/src/lib/controls/textfield/select/select.component.spec.ts
@@ -147,7 +147,7 @@ describe('SelectComponent', () => {
         thenSelectIsNotValuated();
         host.value = '1';
         spectator.detectChanges();
-        tick();
+        tick(100);
         thenSelectHasValue('1', '1');
         expect(spectator.queryAll('pa-dropdown pa-select-options pa-option .pa-option-selected')).toHaveLength(1);
         discardPeriodicTasks();
@@ -189,7 +189,7 @@ describe('SelectComponent', () => {
 
     it('should propagate ngModel value', fakeAsync(() => {
         initWithTemplate(`<pa-select [(ngModel)]="value">${optionsInTemplate}</pa-select>`);
-        tick();
+        tick(100);
         whenFirstOptionClicked();
         thenSelectHasValue('first', 'first label');
         expect(host.value).toEqual('first');
@@ -220,7 +220,7 @@ describe('SelectComponent', () => {
         expect(spectator.query('.pa-select-value')?.innerHTML).toEqual('');
         host.placeholder = 'placeholder';
         spectator.detectChanges();
-        tick();
+        tick(100);
         expect(spectator.query('.pa-select-value')?.innerHTML).toEqual('placeholder');
         whenFirstOptionClicked();
         expect(spectator.query('.pa-select-value')?.innerHTML).toEqual('first label');
@@ -269,7 +269,7 @@ describe('SelectComponent', () => {
         const expanded = jest.spyOn(host, 'onExpanded');
         host.hasFocus = true;
         spectator.detectChanges();
-        tick();
+        tick(100);
         expect(selectClicked).toHaveBeenCalled();
         expect(expanded).toHaveBeenCalledWith(true);
         spectator.detectChanges();
@@ -282,7 +282,7 @@ describe('SelectComponent', () => {
         expect(spectator.query('.pa-select-chevron')).toBeTruthy();
         expect(spectator.query('.pa-select-chevron.opened')).toEqual(null);
         spectator.click('.pa-field-control');
-        tick();
+        tick(100);
         spectator.detectChanges();
         expect(spectator.query('.pa-select-chevron.opened')).toBeTruthy();
         discardPeriodicTasks();
@@ -291,7 +291,7 @@ describe('SelectComponent', () => {
         initWithTemplate(`<pa-select [readonly]="readonly">${optionsInTemplate}</pa-select>`);
         expect((spectator.query<HTMLElement>('.pa-popup') as any).hidden).toEqual(true);
         spectator.click('.pa-select-chevron');
-        tick();
+        tick(100);
         spectator.detectChanges();
         expect((spectator.query<HTMLElement>('.pa-popup') as any).hidden).toEqual(false);
         spectator.click('.pa-select-chevron');
@@ -311,7 +311,7 @@ describe('SelectComponent', () => {
         expect(component.control.pristine).toEqual(true);
         expect(component.control.touched).toEqual(false);
         spectator.click('.pa-field-control');
-        tick();
+        tick(100);
         spectator.detectChanges();
         spectator.triggerEventHandler(DropdownComponent, 'onClose', true);
         expect(component.control.pristine).toEqual(false);
@@ -325,7 +325,7 @@ describe('SelectComponent', () => {
         const selectClicked = jest.spyOn(component.selectInput?.nativeElement, 'click');
         spectator.triggerEventHandler(CdkMonitorFocus, 'cdkFocusChange', 'keyboard');
         spectator.detectChanges();
-        tick();
+        tick(100);
         expect(selectClicked).toHaveBeenCalled();
         spectator.detectChanges();
         expect((spectator.query<HTMLElement>('.pa-popup') as any).hidden).toEqual(false);
@@ -349,7 +349,7 @@ describe('SelectComponent', () => {
         expect(component.control.valid).toEqual(true);
         const statusChanged = jest.spyOn(host, 'statusChanged');
         host.required = true;
-        tick();
+        tick(100);
         spectator.detectChanges();
         expect(component.control.valid).toEqual(false);
         expect(spectator.query('.pa-field-error')).toBeTruthy();
@@ -364,7 +364,7 @@ describe('SelectComponent', () => {
         component.control.markAsDirty();
         expect(component.control.valid).toEqual(true);
         host.required = true;
-        tick();
+        tick(100);
         spectator.detectChanges();
         expect(component.control.valid).toEqual(false);
         const hint = ngMocks.find(spectator.debugElement, FormFieldHintComponent);
@@ -378,7 +378,7 @@ describe('SelectComponent', () => {
 
     it('should render in dim mode', fakeAsync(() => {
         initWithTemplate(`<pa-select label="my field" dim>${optionsInTemplate}</pa-select>`);
-        tick();
+        tick(100);
         expect(spectator.query('.pa-field-container.pa-dim')).toBeTruthy();
         whenFirstOptionClicked();
         expect(spectator.query('label.pa-sr-only')).toBeTruthy();